# gameserver-http/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/gamesession/click:
    post:
      tags:
      - game-session-controller
      operationId: playerClick
      parameters:
      - name: session
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/x-protobuf:
            schema:
              $ref: '#/components/schemas/PlayerClickedEvent'
        required: true
      responses:
        "200":
          description: OK
  /api/gamesession/next:
    get:
      tags:
      - game-session-controller
      operationId: nextTarget
      parameters:
      - name: session
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/x-protobuf:
              schema:
                $ref: '#/components/schemas/ServerGameEvent'
components:
  schemas:
    ByteString:
      type: object
      properties:
        empty:
          type: boolean
        validUtf8:
          type: boolean
    Descriptor:
      type: object
      properties:
        index:
          type: integer
          format: int32
        proto:
          $ref: '#/components/schemas/DescriptorProto'
        fullName:
          type: string
        file:
          $ref: '#/components/schemas/FileDescriptor'
        containingType:
          $ref: '#/components/schemas/Descriptor'
        nestedTypes:
          type: array
          items:
            $ref: '#/components/schemas/Descriptor'
        enumTypes:
          type: array
          items:
            $ref: '#/components/schemas/EnumDescriptor'
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptor'
        extensions:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptor'
        oneofs:
          type: array
          items:
            $ref: '#/components/schemas/OneofDescriptor'
        name:
          type: string
        extendable:
          type: boolean
        options:
          $ref: '#/components/schemas/MessageOptions'
    DescriptorProto:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        name:
          type: string
        nameBytes:
          $ref: '#/components/schemas/ByteString'
        fieldList:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptorProto'
        fieldOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptorProtoOrBuilder'
        extensionList:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptorProto'
        extensionOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptorProtoOrBuilder'
        nestedTypeList:
          type: array
          items:
            $ref: '#/components/schemas/DescriptorProto'
        nestedTypeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/DescriptorProtoOrBuilder'
        enumTypeList:
          type: array
          items:
            $ref: '#/components/schemas/EnumDescriptorProto'
        enumTypeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/EnumDescriptorProtoOrBuilder'
        extensionRangeCount:
          type: integer
          format: int32
        oneofDeclList:
          type: array
          items:
            $ref: '#/components/schemas/OneofDescriptorProto'
        oneofDeclOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/OneofDescriptorProtoOrBuilder'
        optionsOrBuilder:
          $ref: '#/components/schemas/MessageOptionsOrBuilder'
        reservedRangeCount:
          type: integer
          format: int32
        reservedNameCount:
          type: integer
          format: int32
        extensionRangeList:
          type: array
          items:
            $ref: '#/components/schemas/ExtensionRange'
        reservedRangeList:
          type: array
          items:
            $ref: '#/components/schemas/ReservedRange'
        reservedNameList:
          type: array
          properties:
            empty:
              type: boolean
          items:
            type: string
        oneofDeclCount:
          type: integer
          format: int32
        nestedTypeCount:
          type: integer
          format: int32
        enumTypeCount:
          type: integer
          format: int32
        extensionCount:
          type: integer
          format: int32
        fieldCount:
          type: integer
          format: int32
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/DescriptorProto'
        parserForType:
          $ref: '#/components/schemas/ParserDescriptorProto'
        options:
          $ref: '#/components/schemas/MessageOptions'
        extensionRangeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/ExtensionRangeOrBuilder'
        reservedRangeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/ReservedRangeOrBuilder'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true
    DescriptorProtoOrBuilder:
      type: object
      properties:
        name:
          type: string
        nameBytes:
          $ref: '#/components/schemas/ByteString'
        fieldList:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptorProto'
        fieldOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptorProtoOrBuilder'
        extensionList:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptorProto'
        extensionOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptorProtoOrBuilder'
        nestedTypeList:
          type: array
          items:
            $ref: '#/components/schemas/DescriptorProto'
        enumTypeList:
          type: array
          items:
            $ref: '#/components/schemas/EnumDescriptorProto'
        enumTypeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/EnumDescriptorProtoOrBuilder'
        extensionRangeCount:
          type: integer
          format: int32
        oneofDeclList:
          type: array
          items:
            $ref: '#/components/schemas/OneofDescriptorProto'
        oneofDeclOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/OneofDescriptorProtoOrBuilder'
        optionsOrBuilder:
          $ref: '#/components/schemas/MessageOptionsOrBuilder'
        reservedRangeCount:
          type: integer
          format: int32
        reservedNameCount:
          type: integer
          format: int32
        extensionRangeList:
          type: array
          items:
            $ref: '#/components/schemas/ExtensionRange'
        reservedRangeList:
          type: array
          items:
            $ref: '#/components/schemas/ReservedRange'
        reservedNameList:
          type: array
          items:
            type: string
        oneofDeclCount:
          type: integer
          format: int32
        nestedTypeCount:
          type: integer
          format: int32
        enumTypeCount:
          type: integer
          format: int32
        extensionCount:
          type: integer
          format: int32
        fieldCount:
          type: integer
          format: int32
        options:
          $ref: '#/components/schemas/MessageOptions'
        extensionRangeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/ExtensionRangeOrBuilder'
        reservedRangeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/ReservedRangeOrBuilder'
        defaultInstanceForType:
          $ref: '#/components/schemas/Message'
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        initialized:
          type: boolean
    EnumDescriptor:
      type: object
      properties:
        index:
          type: integer
          format: int32
        proto:
          $ref: '#/components/schemas/EnumDescriptorProto'
        fullName:
          type: string
        file:
          $ref: '#/components/schemas/FileDescriptor'
        containingType:
          $ref: '#/components/schemas/Descriptor'
        values:
          type: array
          items:
            $ref: '#/components/schemas/EnumValueDescriptor'
        name:
          type: string
        options:
          $ref: '#/components/schemas/EnumOptions'
    EnumDescriptorProto:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        name:
          type: string
        nameBytes:
          $ref: '#/components/schemas/ByteString'
        optionsOrBuilder:
          $ref: '#/components/schemas/EnumOptionsOrBuilder'
        reservedRangeCount:
          type: integer
          format: int32
        reservedNameCount:
          type: integer
          format: int32
        valueCount:
          type: integer
          format: int32
        reservedRangeList:
          type: array
          items:
            $ref: '#/components/schemas/EnumReservedRange'
        reservedNameList:
          type: array
          properties:
            empty:
              type: boolean
          items:
            type: string
        valueList:
          type: array
          items:
            $ref: '#/components/schemas/EnumValueDescriptorProto'
        valueOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/EnumValueDescriptorProtoOrBuilder'
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/EnumDescriptorProto'
        parserForType:
          $ref: '#/components/schemas/ParserEnumDescriptorProto'
        options:
          $ref: '#/components/schemas/EnumOptions'
        reservedRangeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/EnumReservedRangeOrBuilder'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true
    EnumDescriptorProtoOrBuilder:
      type: object
      properties:
        name:
          type: string
        nameBytes:
          $ref: '#/components/schemas/ByteString'
        optionsOrBuilder:
          $ref: '#/components/schemas/EnumOptionsOrBuilder'
        reservedRangeCount:
          type: integer
          format: int32
        reservedNameCount:
          type: integer
          format: int32
        valueCount:
          type: integer
          format: int32
        reservedRangeList:
          type: array
          items:
            $ref: '#/components/schemas/EnumReservedRange'
        reservedNameList:
          type: array
          items:
            type: string
        valueList:
          type: array
          items:
            $ref: '#/components/schemas/EnumValueDescriptorProto'
        valueOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/EnumValueDescriptorProtoOrBuilder'
        options:
          $ref: '#/components/schemas/EnumOptions'
        reservedRangeOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/EnumReservedRangeOrBuilder'
        defaultInstanceForType:
          $ref: '#/components/schemas/Message'
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        initialized:
          type: boolean
...
    ExtensionRange:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        optionsOrBuilder:
          $ref: '#/components/schemas/ExtensionRangeOptionsOrBuilder'
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/ExtensionRange'
        parserForType:
          $ref: '#/components/schemas/ParserExtensionRange'
        start:
          type: integer
          format: int32
        end:
          type: integer
          format: int32
        options:
          $ref: '#/components/schemas/ExtensionRangeOptions'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true
...
    ExtensionRangeOrBuilder:
      type: object
      properties:
        optionsOrBuilder:
          $ref: '#/components/schemas/ExtensionRangeOptionsOrBuilder'
        start:
          type: integer
          format: int32
        end:
          type: integer
          format: int32
        options:
          $ref: '#/components/schemas/ExtensionRangeOptions'
        defaultInstanceForType:
          $ref: '#/components/schemas/Message'
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        initialized:
          type: boolean
    FieldDescriptor:
      type: object
      properties:
        index:
          type: integer
          format: int32
        proto:
          $ref: '#/components/schemas/FieldDescriptorProto'
        fullName:
          type: string
        jsonName:
          type: string
        file:
          $ref: '#/components/schemas/FileDescriptor'
        extensionScope:
          $ref: '#/components/schemas/Descriptor'
        type:
          type: string
          enum:
          - DOUBLE
          - FLOAT
          - INT64
          - UINT64
          - INT32
          - FIXED64
          - FIXED32
          - BOOL
          - STRING
          - GROUP
          - MESSAGE
          - BYTES
          - UINT32
          - ENUM
          - SFIXED32
          - SFIXED64
          - SINT32
          - SINT64
        containingType:
          $ref: '#/components/schemas/Descriptor'
        messageType:
          $ref: '#/components/schemas/Descriptor'
        containingOneof:
          $ref: '#/components/schemas/OneofDescriptor'
        enumType:
          $ref: '#/components/schemas/EnumDescriptor'
        defaultValue:
          type: object
        name:
          type: string
        number:
          type: integer
          format: int32
        extension:
          type: boolean
        liteType:
          type: string
          enum:
          - DOUBLE
          - FLOAT
          - INT64
          - UINT64
          - INT32
          - FIXED64
          - FIXED32
          - BOOL
          - STRING
          - GROUP
          - MESSAGE
          - BYTES
          - UINT32
          - ENUM
          - SFIXED32
          - SFIXED64
          - SINT32
          - SINT64
        liteJavaType:
          type: string
          enum:
          - INT
          - LONG
          - FLOAT
          - DOUBLE
          - BOOLEAN
          - STRING
          - BYTE_STRING
          - ENUM
          - MESSAGE
        javaType:
          type: string
          enum:
          - INT
          - LONG
          - FLOAT
          - DOUBLE
          - BOOLEAN
          - STRING
          - BYTE_STRING
          - ENUM
          - MESSAGE
        repeated:
          type: boolean
        mapField:
          type: boolean
        options:
          $ref: '#/components/schemas/FieldOptions'
        optional:
          type: boolean
        required:
          type: boolean
        packed:
          type: boolean
        packable:
          type: boolean
    FieldDescriptorProto:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        name:
          type: string
        typeName:
          type: string
        type:
          type: string
          enum:
          - TYPE_DOUBLE
          - TYPE_FLOAT
          - TYPE_INT64
          - TYPE_UINT64
          - TYPE_INT32
          - TYPE_FIXED64
          - TYPE_FIXED32
          - TYPE_BOOL
          - TYPE_STRING
          - TYPE_GROUP
          - TYPE_MESSAGE
          - TYPE_BYTES
          - TYPE_UINT32
          - TYPE_ENUM
          - TYPE_SFIXED32
          - TYPE_SFIXED64
          - TYPE_SINT32
          - TYPE_SINT64
        defaultValue:
          type: string
        number:
          type: integer
          format: int32
        jsonName:
          type: string
        oneofIndex:
          type: integer
          format: int32
        extendee:
          type: string
        nameBytes:
          $ref: '#/components/schemas/ByteString'
        optionsOrBuilder:
          $ref: '#/components/schemas/FieldOptionsOrBuilder'
        typeNameBytes:
          $ref: '#/components/schemas/ByteString'
        extendeeBytes:
          $ref: '#/components/schemas/ByteString'
        defaultValueBytes:
          $ref: '#/components/schemas/ByteString'
        jsonNameBytes:
          $ref: '#/components/schemas/ByteString'
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/FieldDescriptorProto'
        parserForType:
          $ref: '#/components/schemas/ParserFieldDescriptorProto'
        options:
          $ref: '#/components/schemas/FieldOptions'
        label:
          type: string
          enum:
          - LABEL_OPTIONAL
          - LABEL_REQUIRED
          - LABEL_REPEATED
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true
    FieldDescriptorProtoOrBuilder:
      type: object
      properties:
        name:
          type: string
        typeName:
          type: string
        type:
          type: string
          enum:
          - TYPE_DOUBLE
          - TYPE_FLOAT
          - TYPE_INT64
          - TYPE_UINT64
          - TYPE_INT32
          - TYPE_FIXED64
          - TYPE_FIXED32
          - TYPE_BOOL
          - TYPE_STRING
          - TYPE_GROUP
          - TYPE_MESSAGE
          - TYPE_BYTES
          - TYPE_UINT32
          - TYPE_ENUM
          - TYPE_SFIXED32
          - TYPE_SFIXED64
          - TYPE_SINT32
          - TYPE_SINT64
        defaultValue:
          type: string
        number:
          type: integer
          format: int32
        jsonName:
          type: string
        oneofIndex:
          type: integer
          format: int32
        extendee:
          type: string
        nameBytes:
          $ref: '#/components/schemas/ByteString'
        optionsOrBuilder:
          $ref: '#/components/schemas/FieldOptionsOrBuilder'
        typeNameBytes:
          $ref: '#/components/schemas/ByteString'
        extendeeBytes:
          $ref: '#/components/schemas/ByteString'
        defaultValueBytes:
          $ref: '#/components/schemas/ByteString'
        jsonNameBytes:
          $ref: '#/components/schemas/ByteString'
        options:
          $ref: '#/components/schemas/FieldOptions'
        label:
          type: string
          enum:
          - LABEL_OPTIONAL
          - LABEL_REQUIRED
          - LABEL_REPEATED
        defaultInstanceForType:
          $ref: '#/components/schemas/Message'
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        initialized:
          type: boolean
...
    FileDescriptor:
      type: object
      properties:
        proto:
          $ref: '#/components/schemas/FileDescriptorProto'
        messageTypes:
          type: array
          items:
            $ref: '#/components/schemas/Descriptor'
        enumTypes:
          type: array
          items:
            $ref: '#/components/schemas/EnumDescriptor'
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceDescriptor'
        extensions:
          type: array
          items:
            $ref: '#/components/schemas/FieldDescriptor'
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/FileDescriptor'
        publicDependencies:
          type: array
          items:
            $ref: '#/components/schemas/FileDescriptor'
        name:
          type: string
        package:
          type: string
        file:
          $ref: '#/components/schemas/FileDescriptor'
        fullName:
          type: string
        options:
          $ref: '#/components/schemas/FileOptions'
        syntax:
          type: string
          enum:
          - UNKNOWN
          - PROTO2
          - PROTO3
...
    MessageOptions:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        messageSetWireFormat:
          type: boolean
        mapEntry:
          type: boolean
        uninterpretedOptionList:
          type: array
          items:
            $ref: '#/components/schemas/UninterpretedOption'
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/MessageOptions'
        parserForType:
          $ref: '#/components/schemas/ParserMessageOptions'
        deprecated:
          type: boolean
        noStandardDescriptorAccessor:
          type: boolean
        uninterpretedOptionOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/UninterpretedOptionOrBuilder'
        uninterpretedOptionCount:
          type: integer
          format: int32
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        allFieldsRaw:
          type: object
          additionalProperties:
            type: object
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true
    MessageOptionsOrBuilder:
      type: object
      properties:
        messageSetWireFormat:
          type: boolean
        mapEntry:
          type: boolean
        uninterpretedOptionList:
          type: array
          items:
            $ref: '#/components/schemas/UninterpretedOption'
        deprecated:
          type: boolean
        noStandardDescriptorAccessor:
          type: boolean
        uninterpretedOptionOrBuilderList:
          type: array
          items:
            $ref: '#/components/schemas/UninterpretedOptionOrBuilder'
        uninterpretedOptionCount:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/Message'
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        initialized:
          type: boolean
...
    OneofDescriptor:
      type: object
      properties:
        index:
          type: integer
          format: int32
        proto:
          $ref: '#/components/schemas/OneofDescriptorProto'
        fullName:
          type: string
        file:
          $ref: '#/components/schemas/FileDescriptor'
        containingType:
          $ref: '#/components/schemas/Descriptor'
        fieldCount:
          type: integer
          format: int32
        name:
          type: string
        options:
          $ref: '#/components/schemas/OneofOptions'
    OneofDescriptorProto:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        name:
          type: string
        nameBytes:
          $ref: '#/components/schemas/ByteString'
        optionsOrBuilder:
          $ref: '#/components/schemas/OneofOptionsOrBuilder'
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/OneofDescriptorProto'
        parserForType:
          $ref: '#/components/schemas/ParserOneofDescriptorProto'
        options:
          $ref: '#/components/schemas/OneofOptions'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true
    OneofDescriptorProtoOrBuilder:
      type: object
      properties:
        name:
          type: string
        nameBytes:
          $ref: '#/components/schemas/ByteString'
        optionsOrBuilder:
          $ref: '#/components/schemas/OneofOptionsOrBuilder'
        options:
          $ref: '#/components/schemas/OneofOptions'
        defaultInstanceForType:
          $ref: '#/components/schemas/Message'
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        initialized:
          type: boolean
...
    Parser:
      type: object
    ParserDescriptorProto:
      type: object
    ParserEnumDescriptorProto:
      type: object
...
    ParserExtensionRange:
      type: object
...
    ParserFieldDescriptorProto:
      type: object
...
    ParserMessageOptions:
      type: object
...
    ParserOneofDescriptorProto:
      type: object
...
    ParserPlayerClickedEvent:
      type: object
    ParserReservedRange:
      type: object
...
    PlayerClickedEvent:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/PlayerClickedEvent'
        parserForType:
          $ref: '#/components/schemas/ParserPlayerClickedEvent'
        x:
          type: integer
          format: int32
        "y":
          type: integer
          format: int32
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true
    ReservedRange:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/ReservedRange'
        parserForType:
          $ref: '#/components/schemas/ParserReservedRange'
        start:
          type: integer
          format: int32
        end:
          type: integer
          format: int32
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true
    ReservedRangeOrBuilder:
      type: object
      properties:
        start:
          type: integer
          format: int32
        end:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/Message'
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        initialized:
          type: boolean
    ServiceDescriptor:
      type: object
      properties:
        index:
          type: integer
          format: int32
        proto:
          $ref: '#/components/schemas/ServiceDescriptorProto'
        fullName:
          type: string
        file:
          $ref: '#/components/schemas/FileDescriptor'
        methods:
          type: array
          items:
            $ref: '#/components/schemas/MethodDescriptor'
        name:
          type: string
        options:
          $ref: '#/components/schemas/ServiceOptions'
...
    UnknownFieldSet:
      type: object
      properties:
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/UnknownFieldSet'
        parserForType:
          $ref: '#/components/schemas/Parser'
        serializedSizeAsMessageSet:
          type: integer
          format: int32
    ParserServerGameEvent:
      type: object
    ServerGameEvent:
      type: object
      properties:
        unknownFields:
          $ref: '#/components/schemas/UnknownFieldSet'
        initialized:
          type: boolean
        serializedSize:
          type: integer
          format: int32
        defaultInstanceForType:
          $ref: '#/components/schemas/ServerGameEvent'
        parserForType:
          $ref: '#/components/schemas/ParserServerGameEvent'
        score:
          type: integer
          format: int32
        targetX:
          type: integer
          format: int32
        targetY:
          type: integer
          format: int32
        end:
          type: boolean
        allFields:
          type: object
          additionalProperties:
            type: object
        descriptorForType:
          $ref: '#/components/schemas/Descriptor'
        initializationErrorString:
          type: string
        memoizedSerializedSize:
          type: integer
          format: int32
          writeOnly: true